+++ src/main/java/com/supermartijn642/simplemagnets/MagnetItem.java
@@ -45,7 +45,7 @@ public abstract class MagnetItem extends BaseItem {
     }
 
     public static void toggleMagnet(Player player, ItemStack stack){
-        if(!player.level().isClientSide && stack.getItem() instanceof MagnetItem){
+        if(!player.level().isClientSide() && stack.getItem() instanceof MagnetItem){
             //noinspection DataFlowIssue
             boolean active = stack.has(ACTIVE) && stack.get(ACTIVE);
             stack.set(ACTIVE, !active);
@@ -69,17 +69,17 @@ public abstract class MagnetItem extends BaseItem {
                 // TODO look for alternative to 'PreventRemoteMovement' tag on Fabric
 
                 List<ItemEntity> items = level.getEntities(EntityType.ITEM, area,
-                    item -> item.isAlive() && (!level.isClientSide || item.tickCount > 1) &&
+                    item -> item.isAlive() && (!level.isClientSide() || item.tickCount > 1) &&
                         (item.thrower == null || !item.thrower.matches(entity) || !item.hasPickUpDelay()) &&
                         !item.getItem().isEmpty() && ((SimpleMagnetsItemEntity)item).simplemagnetsCanBePickedUp() && this.canPickupStack(stack, item.getItem())
                 );
                 items.forEach(item -> item.setPos(entity.getX(), entity.getY(), entity.getZ()));
                 // Directly add items to the player's inventory when ItemPhysic is installed
-                if(!level.isClientSide && entity instanceof Player && CommonUtils.isModLoaded("itemphysic"))
+                if(!level.isClientSide() && entity instanceof Player && CommonUtils.isModLoaded("itemphysic"))
                     items.forEach(item -> playerTouch(item, (Player)entity));
             }
 
-            if(!level.isClientSide && this.canPickupXp(stack) && entity instanceof Player){
+            if(!level.isClientSide() && this.canPickupXp(stack) && entity instanceof Player){
                 int r = this.getRangeXp(stack);
                 AABB area = new AABB(entity.position().add(-r, -r, -r), entity.position().add(r, r, r));
 
@@ -98,7 +98,7 @@ public abstract class MagnetItem extends BaseItem {
      * Copied from {@link ItemEntity#playerTouch(Player)}. Use this when ItemPhysic is installed to still pick up items.
      */
     private static void playerTouch(ItemEntity itemEntity, Player player){
-        if(!itemEntity.level().isClientSide){
+        if(!itemEntity.level().isClientSide()){
             ItemStack itemstack = itemEntity.getItem();
             Item item = itemstack.getItem();
             int i = itemstack.getCount();
@@ -140,7 +140,7 @@ public abstract class MagnetItem extends BaseItem {
     protected abstract Component getTooltip();
 
     public static void onStartTracking(Entity target, Player player){
-        if(!player.level().isClientSide && target instanceof ItemEntity && ((ItemEntity)target).thrower != null)
+        if(!player.level().isClientSide() && target instanceof ItemEntity && ((ItemEntity)target).thrower != null)
             SimpleMagnets.CHANNEL.sendToPlayer(player, new PacketItemInfo((ItemEntity)target));
     }
 }